revert(#4): fix(#4): friends was fixed
This reverts '5f7747c28941e711f4a207f08e21f74a8fc798af'

+++ database/src/main/kotlin/app/meetacy/backend/database/notifications/NotificationsTable.kt
@@ -46,10 +46,10 @@ class NotificationsTable(private val db: Database) : Table() {
 
     }
 
-    fun getNotifications(ownerId: UserId, offset: Int, amount: Int): List<DatabaseNotification> =
+    fun getNotifications(ownerId: UserId, offset: Long, amount: Int): List<DatabaseNotification> =
         transaction(db) {
             select { (OWNER_ID eq ownerId.long) }
-                .limit(amount, offset.toLong())
+                .limit(amount, offset)
                 .map { it ->
                     DatabaseNotification(
                         NotificationId(it[NOTIFICATION_ID]),
